// apps/site/src/ui/Dashboard/BillingShipping/type.ts
// Dashboard Billing Shipping multi language text type
export type textData = {
	billingText: string;
	shippingText: string;
	editText: string;
	noResultText: string;
	fromText: {
		firstNamePlaceholder?: string
		lastNamePlaceholder: string;
		companyPlaceholder: string;
		emailPlaceholder: string;
		phonePlaceholder: string;
		countryPlaceholder: string;
		address1Placeholder: string;
		address2Placeholder: string;
		statePlaceholder: string;
		cityPlaceholder: string;
		postCodePlaceholder: string;
		buttonText: string;
	}
}